requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
paramiko>=3.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
